+++ fourtooncookie/src/test/java/com/startingblue/fourtooncookie/global/authorization/member/MemberSignedUpAuthorizationInterceptorTest.java
@@ -1,6 +1,5 @@
-package com.startingblue.fourtooncookie.member.authorization;
+package com.startingblue.fourtooncookie.global.authorization.member;
 
-import com.startingblue.fourtooncookie.global.authorization.member.MemberSignedUpAuthorizationInterceptor;
 import com.startingblue.fourtooncookie.member.service.MemberService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
